refactor(control-api): split tenants.rs into submodules to clear 600-line cap
